+++ service/cloudwatch/doc.go
@@ -3,10 +3,10 @@
 // Package cloudwatch provides the client and types for making API
 // requests to Amazon CloudWatch.
 //
-// Amazon CloudWatch monitors your Amazon Web Services (AWS) resources and the
-// applications you run on AWS in real time. You can use CloudWatch to collect
-// and track metrics, which are the variables you want to measure for your resources
-// and applications.
+// Amazon CloudWatch monitors your Amazon Web Services (Amazon Web Services)
+// resources and the applications you run on Amazon Web Services in real time.
+// You can use CloudWatch to collect and track metrics, which are the variables
+// you want to measure for your resources and applications.
 //
 // CloudWatch alarms send notifications or automatically change the resources
 // you are monitoring based on rules that you define. For example, you can monitor
@@ -15,9 +15,10 @@
 // to handle increased load. You can also use this data to stop under-used instances
 // to save money.
 //
-// In addition to monitoring the built-in metrics that come with AWS, you can
-// monitor your own custom metrics. With CloudWatch, you gain system-wide visibility
-// into resource utilization, application performance, and operational health.
+// In addition to monitoring the built-in metrics that come with Amazon Web
+// Services, you can monitor your own custom metrics. With CloudWatch, you gain
+// system-wide visibility into resource utilization, application performance,
+// and operational health.
 //
 // See https://docs.aws.amazon.com/goto/WebAPI/monitoring-2010-08-01 for more information on this service.
 //