+++ service/identitystore/doc.go
@@ -3,6 +3,10 @@
 // Package identitystore provides the client and types for making API
 // requests to AWS SSO Identity Store.
 //
+// The AWS Single Sign-On (SSO) Identity Store service provides a single place
+// to retrieve all of your identities (users and groups). For more information
+// about AWS, see the AWS Single Sign-On User Guide (https://docs.aws.amazon.com/singlesignon/latest/userguide/what-is.html).
+//
 // See https://docs.aws.amazon.com/goto/WebAPI/identitystore-2020-06-15 for more information on this service.
 //
 // See identitystore package documentation for more information.