+++ service/textract/api.go
@@ -161,6 +161,132 @@ func (c *Textract) AnalyzeDocumentWithContext(ctx aws.Context, input *AnalyzeDoc
 	return out, req.Send()
 }
 
+const opAnalyzeExpense = "AnalyzeExpense"
+
+// AnalyzeExpenseRequest generates a "aws/request.Request" representing the
+// client's request for the AnalyzeExpense operation. The "output" return
+// value will be populated with the request's response once the request completes
+// successfully.
+//
+// Use "Send" method on the returned Request to send the API call to the service.
+// the "output" return value is not valid until after Send returns without error.
+//
+// See AnalyzeExpense for more information on using the AnalyzeExpense
+// API call, and error handling.
+//
+// This method is useful when you want to inject custom logic or configuration
+// into the SDK's request lifecycle. Such as custom headers, or retry logic.
+//
+//
+//    // Example sending a request using the AnalyzeExpenseRequest method.
+//    req, resp := client.AnalyzeExpenseRequest(params)
+//
+//    err := req.Send()
+//    if err == nil { // resp is now filled
+//        fmt.Println(resp)
+//    }
+//
+// See also, https://docs.aws.amazon.com/goto/WebAPI/textract-2018-06-27/AnalyzeExpense
+func (c *Textract) AnalyzeExpenseRequest(input *AnalyzeExpenseInput) (req *request.Request, output *AnalyzeExpenseOutput) {
+	op := &request.Operation{
+		Name:       opAnalyzeExpense,
+		HTTPMethod: "POST",
+		HTTPPath:   "/",
+	}
+
+	if input == nil {
+		input = &AnalyzeExpenseInput{}
+	}
+
+	output = &AnalyzeExpenseOutput{}
+	req = c.newRequest(op, input, output)
+	return
+}
+
+// AnalyzeExpense API operation for Amazon Textract.
+//
+// Analyzes an input document for financially related relationships between
+// text.
+//
+// Information is returned as ExpenseDocuments and seperated as follows.
+//
+//    * LineItemGroups- A data set containing LineItems which store information
+//    about the lines of text, such as an item purchased and its price on a
+//    receipt.
+//
+//    * SummaryFields- Contains all other information a receipt, such as header
+//    information or the vendors name.
+//
+// Returns awserr.Error for service API and SDK errors. Use runtime type assertions
+// with awserr.Error's Code and Message methods to get detailed information about
+// the error.
+//
+// See the AWS API reference guide for Amazon Textract's
+// API operation AnalyzeExpense for usage and error information.
+//
+// Returned Error Types:
+//   * InvalidParameterException
+//   An input parameter violated a constraint. For example, in synchronous operations,
+//   an InvalidParameterException exception occurs when neither of the S3Object
+//   or Bytes values are supplied in the Document request parameter. Validate
+//   your parameter before calling the API operation again.
+//
+//   * InvalidS3ObjectException
+//   Amazon Textract is unable to access the S3 object that's specified in the
+//   request. for more information, Configure Access to Amazon S3 (https://docs.aws.amazon.com/AmazonS3/latest/dev/s3-access-control.html)
+//   For troubleshooting information, see Troubleshooting Amazon S3 (https://docs.aws.amazon.com/AmazonS3/latest/dev/troubleshooting.html)
+//
+//   * UnsupportedDocumentException
+//   The format of the input document isn't supported. Documents for synchronous
+//   operations can be in PNG or JPEG format. Documents for asynchronous operations
+//   can also be in PDF format.
+//
+//   * DocumentTooLargeException
+//   The document can't be processed because it's too large. The maximum document
+//   size for synchronous operations 10 MB. The maximum document size for asynchronous
+//   operations is 500 MB for PDF files.
+//
+//   * BadDocumentException
+//   Amazon Textract isn't able to read the document. For more information on
+//   the document limits in Amazon Textract, see limits.
+//
+//   * AccessDeniedException
+//   You aren't authorized to perform the action. Use the Amazon Resource Name
+//   (ARN) of an authorized user or IAM role to perform the operation.
+//
+//   * ProvisionedThroughputExceededException
+//   The number of requests exceeded your throughput limit. If you want to increase
+//   this limit, contact Amazon Textract.
+//
+//   * InternalServerError
+//   Amazon Textract experienced a service issue. Try your call again.
+//
+//   * ThrottlingException
+//   Amazon Textract is temporarily unable to process the request. Try your call
+//   again.
+//
+// See also, https://docs.aws.amazon.com/goto/WebAPI/textract-2018-06-27/AnalyzeExpense
+func (c *Textract) AnalyzeExpense(input *AnalyzeExpenseInput) (*AnalyzeExpenseOutput, error) {
+	req, out := c.AnalyzeExpenseRequest(input)
+	return out, req.Send()
+}
+
+// AnalyzeExpenseWithContext is the same as AnalyzeExpense with the addition of
+// the ability to pass a context and additional request options.
+//
+// See AnalyzeExpense for details on how to use this API operation.
+//
+// The context must be non-nil and will be used for request cancellation. If
+// the context is nil a panic will occur. In the future the SDK may create
+// sub-contexts for http.Requests. See https://golang.org/pkg/context/
+// for more information on using Contexts.
+func (c *Textract) AnalyzeExpenseWithContext(ctx aws.Context, input *AnalyzeExpenseInput, opts ...request.Option) (*AnalyzeExpenseOutput, error) {
+	req, out := c.AnalyzeExpenseRequest(input)
+	req.SetContext(ctx)
+	req.ApplyOptions(opts...)
+	return out, req.Send()
+}
+
 const opDetectDocumentText = "DetectDocumentText"
 
 // DetectDocumentTextRequest generates a "aws/request.Request" representing the
@@ -413,6 +539,10 @@ func (c *Textract) GetDocumentAnalysisRequest(input *GetDocumentAnalysisInput) (
 //   request. for more information, Configure Access to Amazon S3 (https://docs.aws.amazon.com/AmazonS3/latest/dev/s3-access-control.html)
 //   For troubleshooting information, see Troubleshooting Amazon S3 (https://docs.aws.amazon.com/AmazonS3/latest/dev/troubleshooting.html)
 //
+//   * InvalidKMSKeyException
+//   Indicates you do not have decrypt permissions with the KMS key entered, or
+//   the KMS key was entered incorrectly.
+//
 // See also, https://docs.aws.amazon.com/goto/WebAPI/textract-2018-06-27/GetDocumentAnalysis
 func (c *Textract) GetDocumentAnalysis(input *GetDocumentAnalysisInput) (*GetDocumentAnalysisOutput, error) {
 	req, out := c.GetDocumentAnalysisRequest(input)
@@ -544,6 +674,10 @@ func (c *Textract) GetDocumentTextDetectionRequest(input *GetDocumentTextDetecti
 //   Amazon Textract is unable to access the S3 object that's specified in the
 //   request. for more information, Configure Access to Amazon S3 (https://docs.aws.amazon.com/AmazonS3/latest/dev/s3-access-control.html)
 //   For troubleshooting information, see Troubleshooting Amazon S3 (https://docs.aws.amazon.com/AmazonS3/latest/dev/troubleshooting.html)
+//
+//   * InvalidKMSKeyException
+//   Indicates you do not have decrypt permissions with the KMS key entered, or
+//   the KMS key was entered incorrectly.
 //
 // See also, https://docs.aws.amazon.com/goto/WebAPI/textract-2018-06-27/GetDocumentTextDetection
 func (c *Textract) GetDocumentTextDetection(input *GetDocumentTextDetectionInput) (*GetDocumentTextDetectionOutput, error) {
@@ -1046,6 +1180,102 @@ func (s *AnalyzeDocumentOutput) SetDocumentMetadata(v *DocumentMetadata) *Analyz
 // SetHumanLoopActivationOutput sets the HumanLoopActivationOutput field's value.
 func (s *AnalyzeDocumentOutput) SetHumanLoopActivationOutput(v *HumanLoopActivationOutput) *AnalyzeDocumentOutput {
 	s.HumanLoopActivationOutput = v
+	return s
+}
+
+type AnalyzeExpenseInput struct {
+	_ struct{} `type:"structure"`
+
+	// The input document, either as bytes or as an S3 object.
+	//
+	// You pass image bytes to an Amazon Textract API operation by using the Bytes
+	// property. For example, you would use the Bytes property to pass a document
+	// loaded from a local file system. Image bytes passed by using the Bytes property
+	// must be base64 encoded. Your code might not need to encode document file
+	// bytes if you're using an AWS SDK to call Amazon Textract API operations.
+	//
+	// You pass images stored in an S3 bucket to an Amazon Textract API operation
+	// by using the S3Object property. Documents stored in an S3 bucket don't need
+	// to be base64 encoded.
+	//
+	// The AWS Region for the S3 bucket that contains the S3 object must match the
+	// AWS Region that you use for Amazon Textract operations.
+	//
+	// If you use the AWS CLI to call Amazon Textract operations, passing image
+	// bytes using the Bytes property isn't supported. You must first upload the
+	// document to an Amazon S3 bucket, and then call the operation using the S3Object
+	// property.
+	//
+	// For Amazon Textract to process an S3 object, the user must have permission
+	// to access the S3 object.
+	//
+	// Document is a required field
+	Document *Document `type:"structure" required:"true"`
+}
+
+// String returns the string representation
+func (s AnalyzeExpenseInput) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s AnalyzeExpenseInput) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *AnalyzeExpenseInput) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "AnalyzeExpenseInput"}
+	if s.Document == nil {
+		invalidParams.Add(request.NewErrParamRequired("Document"))
+	}
+	if s.Document != nil {
+		if err := s.Document.Validate(); err != nil {
+			invalidParams.AddNested("Document", err.(request.ErrInvalidParams))
+		}
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetDocument sets the Document field's value.
+func (s *AnalyzeExpenseInput) SetDocument(v *Document) *AnalyzeExpenseInput {
+	s.Document = v
+	return s
+}
+
+type AnalyzeExpenseOutput struct {
+	_ struct{} `type:"structure"`
+
+	// Information about the input document.
+	DocumentMetadata *DocumentMetadata `type:"structure"`
+
+	// The expenses detected by Amazon Textract.
+	ExpenseDocuments []*ExpenseDocument `type:"list"`
+}
+
+// String returns the string representation
+func (s AnalyzeExpenseOutput) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s AnalyzeExpenseOutput) GoString() string {
+	return s.String()
+}
+
+// SetDocumentMetadata sets the DocumentMetadata field's value.
+func (s *AnalyzeExpenseOutput) SetDocumentMetadata(v *DocumentMetadata) *AnalyzeExpenseOutput {
+	s.DocumentMetadata = v
+	return s
+}
+
+// SetExpenseDocuments sets the ExpenseDocuments field's value.
+func (s *AnalyzeExpenseOutput) SetExpenseDocuments(v []*ExpenseDocument) *AnalyzeExpenseOutput {
+	s.ExpenseDocuments = v
 	return s
 }
 
@@ -1685,6 +1915,180 @@ func (s *DocumentTooLargeException) StatusCode() int {
 // RequestID returns the service's response RequestID for request.
 func (s *DocumentTooLargeException) RequestID() string {
 	return s.RespMetadata.RequestID
+}
+
+// An object used to store information about the Value or Label detected by
+// Amazon Textract.
+type ExpenseDetection struct {
+	_ struct{} `type:"structure"`
+
+	// The confidence in detection, as a percentage
+	Confidence *float64 `type:"float"`
+
+	// Information about where the following items are located on a document page:
+	// detected page, text, key-value pairs, tables, table cells, and selection
+	// elements.
+	Geometry *Geometry `type:"structure"`
+
+	// The word or line of text recognized by Amazon Textract
+	Text *string `type:"string"`
+}
+
+// String returns the string representation
+func (s ExpenseDetection) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s ExpenseDetection) GoString() string {
+	return s.String()
+}
+
+// SetConfidence sets the Confidence field's value.
+func (s *ExpenseDetection) SetConfidence(v float64) *ExpenseDetection {
+	s.Confidence = &v
+	return s
+}
+
+// SetGeometry sets the Geometry field's value.
+func (s *ExpenseDetection) SetGeometry(v *Geometry) *ExpenseDetection {
+	s.Geometry = v
+	return s
+}
+
+// SetText sets the Text field's value.
+func (s *ExpenseDetection) SetText(v string) *ExpenseDetection {
+	s.Text = &v
+	return s
+}
+
+// The structure holding all the information returned by AnalyzeExpense
+type ExpenseDocument struct {
+	_ struct{} `type:"structure"`
+
+	// Denotes which invoice or receipt in the document the information is coming
+	// from. First document will be 1, the second 2, and so on.
+	ExpenseIndex *int64 `type:"integer"`
+
+	// Information detected on each table of a document, seperated into LineItems.
+	LineItemGroups []*LineItemGroup `type:"list"`
+
+	// Any information found outside of a table by Amazon Textract.
+	SummaryFields []*ExpenseField `type:"list"`
+}
+
+// String returns the string representation
+func (s ExpenseDocument) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s ExpenseDocument) GoString() string {
+	return s.String()
+}
+
+// SetExpenseIndex sets the ExpenseIndex field's value.
+func (s *ExpenseDocument) SetExpenseIndex(v int64) *ExpenseDocument {
+	s.ExpenseIndex = &v
+	return s
+}
+
+// SetLineItemGroups sets the LineItemGroups field's value.
+func (s *ExpenseDocument) SetLineItemGroups(v []*LineItemGroup) *ExpenseDocument {
+	s.LineItemGroups = v
+	return s
+}
+
+// SetSummaryFields sets the SummaryFields field's value.
+func (s *ExpenseDocument) SetSummaryFields(v []*ExpenseField) *ExpenseDocument {
+	s.SummaryFields = v
+	return s
+}
+
+// Breakdown of detected information, seperated into the catagories Type, LableDetection,
+// and ValueDetection
+type ExpenseField struct {
+	_ struct{} `type:"structure"`
+
+	// The explicitly stated label of a detected element.
+	LabelDetection *ExpenseDetection `type:"structure"`
+
+	// The page number the value was detected on.
+	PageNumber *int64 `type:"integer"`
+
+	// The implied label of a detected element. Present alongside LabelDetection
+	// for explicit elements.
+	Type *ExpenseType `type:"structure"`
+
+	// The value of a detected element. Present in explicit and implicit elements.
+	ValueDetection *ExpenseDetection `type:"structure"`
+}
+
+// String returns the string representation
+func (s ExpenseField) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s ExpenseField) GoString() string {
+	return s.String()
+}
+
+// SetLabelDetection sets the LabelDetection field's value.
+func (s *ExpenseField) SetLabelDetection(v *ExpenseDetection) *ExpenseField {
+	s.LabelDetection = v
+	return s
+}
+
+// SetPageNumber sets the PageNumber field's value.
+func (s *ExpenseField) SetPageNumber(v int64) *ExpenseField {
+	s.PageNumber = &v
+	return s
+}
+
+// SetType sets the Type field's value.
+func (s *ExpenseField) SetType(v *ExpenseType) *ExpenseField {
+	s.Type = v
+	return s
+}
+
+// SetValueDetection sets the ValueDetection field's value.
+func (s *ExpenseField) SetValueDetection(v *ExpenseDetection) *ExpenseField {
+	s.ValueDetection = v
+	return s
+}
+
+// An object used to store information about the Type detected by Amazon Textract.
+type ExpenseType struct {
+	_ struct{} `type:"structure"`
+
+	// The confidence of accuracy, as a percentage.
+	Confidence *float64 `type:"float"`
+
+	// The word or line of text detected by Amazon Textract.
+	Text *string `type:"string"`
+}
+
+// String returns the string representation
+func (s ExpenseType) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s ExpenseType) GoString() string {
+	return s.String()
+}
+
+// SetConfidence sets the Confidence field's value.
+func (s *ExpenseType) SetConfidence(v float64) *ExpenseType {
+	s.Confidence = &v
+	return s
+}
+
+// SetText sets the Text field's value.
+func (s *ExpenseType) SetText(v string) *ExpenseType {
+	s.Text = &v
+	return s
 }
 
 // Information about where the following items are located on a document page:
@@ -2634,6 +3038,66 @@ func (s *LimitExceededException) RequestID() string {
 	return s.RespMetadata.RequestID
 }
 
+// A structure that holds information about the different lines found in a document's
+// tables.
+type LineItemFields struct {
+	_ struct{} `type:"structure"`
+
+	// ExpenseFields used to show information from detected lines on a table.
+	LineItemExpenseFields []*ExpenseField `type:"list"`
+}
+
+// String returns the string representation
+func (s LineItemFields) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s LineItemFields) GoString() string {
+	return s.String()
+}
+
+// SetLineItemExpenseFields sets the LineItemExpenseFields field's value.
+func (s *LineItemFields) SetLineItemExpenseFields(v []*ExpenseField) *LineItemFields {
+	s.LineItemExpenseFields = v
+	return s
+}
+
+// A grouping of tables which contain LineItems, with each table identified
+// by the table's LineItemGroupIndex.
+type LineItemGroup struct {
+	_ struct{} `type:"structure"`
+
+	// The number used to identify a specific table in a document. The first table
+	// encountered will have a LineItemGroupIndex of 1, the second 2, etc.
+	LineItemGroupIndex *int64 `type:"integer"`
+
+	// The breakdown of information on a particular line of a table.
+	LineItems []*LineItemFields `type:"list"`
+}
+
+// String returns the string representation
+func (s LineItemGroup) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s LineItemGroup) GoString() string {
+	return s.String()
+}
+
+// SetLineItemGroupIndex sets the LineItemGroupIndex field's value.
+func (s *LineItemGroup) SetLineItemGroupIndex(v int64) *LineItemGroup {
+	s.LineItemGroupIndex = &v
+	return s
+}
+
+// SetLineItems sets the LineItems field's value.
+func (s *LineItemGroup) SetLineItems(v []*LineItemFields) *LineItemGroup {
+	s.LineItems = v
+	return s
+}
+
 // The Amazon Simple Notification Service (Amazon SNS) topic to which Amazon
 // Textract publishes the completion status of an asynchronous document operation,
 // such as StartDocumentTextDetection.
@@ -2698,6 +3162,24 @@ func (s *NotificationChannel) SetSNSTopicArn(v string) *NotificationChannel {
 
 // Sets whether or not your output will go to a user created bucket. Used to
 // set the name of the bucket, and the prefix on the output file.
+//
+// OutputConfig is an optional parameter which lets you adjust where your output
+// will be placed. By default, Amazon Textract will store the results internally
+// and can only be accessed by the Get API operations. With OutputConfig enabled,
+// you can set the name of the bucket the output will be sent to and the file
+// prefix of the results where you can download your results. Additionally,
+// you can set the KMSKeyID parameter to a customer master key (CMK) to encrypt
+// your output. Without this parameter set Amazon Textract will encrypt server-side
+// using the AWS managed CMK for Amazon S3.
+//
+// Decryption of Customer Content is necessary for processing of the documents
+// by Amazon Textract. If your account is opted out under an AI services opt
+// out policy then all unencrypted Customer Content is immediately and permanently
+// deleted after the Customer Content has been processed by the service. No
+// copy of of the output is retained by Amazon Textract. For information about
+// how to opt out, see Managing AI services opt-out policy. (https://docs.aws.amazon.com/organizations/latest/userguide/orgs_manage_policies_ai-opt-out.html)
+//
+// For more information on data privacy, see the Data Privacy FAQ (https://aws.amazon.com/compliance/data-privacy-faq/).
 type OutputConfig struct {
 	_ struct{} `type:"structure"`
 
@@ -2903,7 +3385,8 @@ func (s *Relationship) SetType(v string) *Relationship {
 type S3Object struct {
 	_ struct{} `type:"structure"`
 
-	// The name of the S3 bucket.
+	// The name of the S3 bucket. Note that the # character is not valid in the
+	// file name.
 	Bucket *string `min:"3" type:"string"`
 
 	// The file name of the input document. Synchronous operations can use image